cleaned unused code
