+++ src/index.html
@@ -113,10 +113,6 @@
 
 </header>
 
-<div id="container3">
-  <button id="random-button">Random</button>
-  <div id="random-course"></div>
-</div>
 <div id="container2">
   <div class="buttons">
     <button id="restaurant-sodexo">Myyrmäki</button>
@@ -126,7 +122,6 @@
   </div>
   <div class="buttons">
     <button id="language-button">Suomi</button>
-    <button id="gluten-button">Gluteeniton</button>
     <button id="darkmode-button">Darkmode</button>
   </div>
 </div>
