+++ src/modules/restaurant-info.js
@@ -1,0 +1,32 @@
+const restaurants = [
+  {
+    name: 'Myyrmaki',
+    id: 152,
+    type: 'sodexo',
+    lon: '24.844454601853037',
+    lat: '60.258748360347994',
+  },
+  {
+    name: 'Karamalmi',
+    id: 3208,
+    type: 'fazer',
+    lon: '24.758666255819254',
+    lat: '60.22419827645741',
+  },
+  {
+    name: 'Myllypuro',
+    id: 158,
+    type: 'sodexo',
+    lon: '25.0779246',
+    lat: '60.223573821912986',
+  },
+  {
+    name: 'Arabia',
+    id: 1251,
+    type: 'fazer',
+    lon: '24.97673291349155',
+    lat: '60.21003284593111',
+  },
+];
+
+export {restaurants};
